+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-01-09
+* Added `EnableLocalPlayerLobbyEvents` to enable/disable Lobby events triggered by the local player
+
 ## [1.1.2] - 2023-12-06
 * Fixed Lobby Vivox interoperability issues around joining certain channel types or joining channels that didn't match a Lobby ID when trying only to use the Vivox SDK while the Lobby SDK was present in the project.
 
@@ -82,7 +85,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-* Introduced LobbyEvents functionality to allow receiving runtime updates when used with the Wire package.
+* Introduced LobbyEvents functionality behind a feature flag to allow receiving runtime updates when used with the Wire package.
 * Replaced "Lobbies" and "ILobbyServiceSDK", replaced with "LobbyService" and "ILobbyService"
 * Updating API Base Path.